requirements.txt
# Appium
appium-python-client>=4.0.0

# Testing framework
pytest>=8.0.0
pytest-xdist>=3.5.0
pytest-rerunfailures>=14.0
pytest-html>=4.1.0

# Reporting
allure-pytest>=2.13.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# GPS/GPX handling
gpxpy>=1.6.2

# HTTP client
requests>=2.31.0

# Utilities
python-dateutil>=2.8.2

# Visual regression testing
Pillow>=10.0.0
scikit-image>=0.22.0
numpy>=1.24.0
